Move people components
