+++ src/components/project/ProjectPage.tsx
@@ -4,7 +4,7 @@ import { resolve } from 'ketting';
 import { useNavigate, useLocation } from 'react-router-dom';
 import { Project } from '@badgateway/tt-types';
 
-import { ClientSelect } from '../client/ClientSelect';
+import { ClientSelect } from '../ClientSelect';
 
 export function ProjectPage() {
